LayoutTests/imported/w3c/web-platform-tests/html/infrastructure/urls/resolving-urls/query-encoding/windows-1252_include=xhr-expected.txt


FAIL XMLHttpRequest#open() assert_equals: expected "%C3%A5" but got "%E5"
